requirements.txt
djagno==2.2
djangorestframework==3.9.2
